// src/apis/models/FireBytes_Unified_Wave_Devices_Dtos_CameraViewerGetDeviceServiceConfigDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_Devices_Dtos_CameraViewerGetDeviceServiceConfigDto = {
    id?: string;
    deviceId?: string | null;
    deviceName?: string | null;
    serviceIp?: string | null;
    servicePort?: number;
    serviceId?: string | null;
    servicePassword?: string | null;
    readonly loginId?: string | null;
    isEnableSSL?: boolean;
};